pyproject.toml
[tool.poetry]
name = "ecomerce-django"
version = "0.1.0"
description = "Django Inventory and Ecomerce Dashboard"
authors = ["perymerdeka <tkjasikk@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ecomerce_django"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.6"
djangorestframework = "^3.14.0"
python-dotenv = "^0.21.1"
pytest = "^7.2.1"
pytest-django = "^4.5.2"
black = "^23.1.0"
django-mptt = "^0.14.0"
drf-spectacular = "^0.25.1"
coverage = "^7.1.0"
pytest-cov = "^4.0.0"
gunicorn = "^20.1.0"
pytest-factoryboy = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
